Another fix from Olly, via the Xapian sources, to analyser.c: Cast the arguments, not the function pointer as ISO C allows `void*' and `struct amongvec*' to have different representations.
Also, modify the function prototypes of the comparison functions in
analyser.c, sort.c and header.h to take "const void *" instead of "void *".
This is a good thing since we don't allow the comparison function to modify
the values being compared, but also means that our comparison function
could be passed to "qsort" instead of using the snowball implementation of
sort with just a one-line modification of the sources.  Xapian is now doing
this, and it would be nice to keep our differences from Xapian's sources
minimal.



git-svn-id: svn+userv://snowball.tartarus.org/snowball/trunk/snowball@428 633ccae0-01f4-0310-8c99-d3591da6f01f