Another patch from Olly Betts:
This changes the generated C code to avoid shadowed local variables.  In
particular it generates int c1, c2, c3 instead of int c and the same for
m.  Also, the existing use of m3 is renamed to mlimit (which is clearer
anyway).  The motivations are that this makes the generated code easier
to understand to the human reader, and allows it to be compiled without
warnings with the GCC option -Wshadow, which avoids some special casing
in Xapian's build system.


git-svn-id: svn+userv://snowball.tartarus.org/snowball/trunk/snowball@437 633ccae0-01f4-0310-8c99-d3591da6f01f