Fix functions in among to work in Java
We seem to need to make the methods called from among 'public' instead
of 'private', and to call them on 'this' instead of the 'methodObject'
(which is cleaner anyway).

I failed to find any version under version control where this worked,
but Richard says it definitely used to - possibly older JVMs failed to
correctly enforce the access controls when methods were invoked by
reflection.