Add 'len' and 'lenof' to measure Unicode length
These are similar to 'size' and 'sizeof' (respectively), but 'size' and
'sizeof' return the length in bytes under '-utf8', whereas these new
commands give the same result whether using '-utf8', '-widechars' or
neither (but under '-utf8' they are O(n) in the length of the string).

For compatibility with existing code which might use these as variable
or function names, they stop being treated as tokens if declared to be
a variable or function.