[java] Fix $ on strings
The previous generated code was just wrong.  This doesn't affect
any of the included algorithms, but for example breaks Martin
Porter's snowball implementation of Schinke's Latin Stemmer.

Issue noted by Jakob Demler while working on the Rust backend in #51,
and reported in the Schinke's Latin Stemmer by Alexander Myltsev
in #58.