[ci] Update .travis.yml
Really fix handling of PRs which also update snowball-data.

Don't require sudo as we don't need it.

Split up the multi-language builds for better status visibility.  With
the container images the overhead of each build is lower.

Update the python versions to try to match what travis actually has
(somewhat guessed as that doesn't seem to be clearly documented).