Avoid infinite recursion in compiler
Recursive functions aren't typical in snowball programs, but the
compiler could end up in infinite recursion (typically failing with
a segfault when it ran out of stack) if fed such a program.

To avoid this, we simply put a limit on how deep the compiler will
recurse and make the pessimistic assumption in the unlikely event
we hit this limit.