Clean up default option handling
Set string options to NULL initially, then set any language-specific
default if they are still NULL after handling command-line options.
This avoids odd behaviour such as using the C# defaults with the Java
backend if the user specifies -csharp -java (which is a bit of an
odd thing to do, but works otherwise).  Handle encoding options in
a different but with a similar end result, plus warning if they
are used for a language with a fixed encoding.