Use more sensible initial input buffer size
Snowball programs are typically a few KB in size (with the current
largest we ship being the Arabic stemmer at 17.1KB) so starting with
a 10 byte input buffer and increasing its size by 50% plus 40 bytes
each time it fills is inefficient - it needs up to 14 reallocations to
load the sources for the stemmers we ship.

Instead start at 8192 bytes and double the size each time we fill it.