Remove "memory leak" entry
This was only in the snowball compiler, not the generated code for
Python (or any other language).  And it wasn't really a memory "leak"
since the compiler exits right after this memory is released at
which point the OS will reclaim all the memory from the process - this
change really just gets rid of a message about "unfreed blocks".