Store booleans in same array as integers
This means each boolean is stored as an int instead of an unsigned char
which means 4 bytes instead of 1, but we save a pointer (4 or 8 bytes)
in struct SN_env so that's a win for all the current stemmers.

For an algorithm which uses both integers and booleans, we also save the
overhead of allocating a block on the heap, and potentially improve
data locality.