build: fix bug about GNUmakefile dependencies on Windows
On Windows platforms executables have the extension '.exe'.  Compilers
like gcc append the extension if not provided in the '-o' flag.  A Windows
platform can easily be recognized by checking if the environment variable
OS equals 'Windows_NT'.

This commit adds the extension of built executables to prerequisites of
GNUmakefile rules, such that make can figure out the dependencies.

NOTE: Extensions are not added to executable names when called by a shell
(e.g., cmd, sh) which are usually able to call the executable even when
the name is provided without extension.