ci: Add "distribution builds"
These do "make dist" and then unpack and build the C libstemmer.

We now only apply the C90 checks to these builds.  Having the generated
code and runtime support work with C90 may still be useful for some
users, but it's reasonable to require a C99 compiler to build the
snowball compiler, and it would be good to be able to use C99 features
which make for clearer and more maintainable code.