C: Handle negative hop as documented
The manual says a negative argument to hop will raise signal f, but
the implementation for all languages is actually to move the cursor
in the opposite direction.  The implemented behaviour is problematic
as it allows invalidating saved cursor values by modifying the string
outside the current region, and only the new yiddish algorithm makes
use of negative hop with the implemented semantics, so we've decided
it's best to fix the implementation to match the documentation.