python: End if-chain with else where possible
In an `among` where all cases have commands we don't need to test
the value of `among_var` in the final case.

This change seems to speed up `make check_python_arabic` by a few
percent (I was getting quite a lot of variation in my tests, so
hard to give a precise number).