yiddish: Fix removal of -ndik suffix
A bug in the code meant we'd always remove -endik, with the code which
was meant to decide whether or not to leave the Ayen in place not
actually having any effect.  That meant a second bug in this code
(it checks for the Latin alphabet 5-character string "Lamed" instead
of the Hebrew alphabet character Lamed) had no effect.

This fix affects 1230 words in the test vocabulary (~1.2%).