yiddish: Eliminate uses of hop with a negative arg
The Snowball Manual has always documented this case as failing with
signal f, but the implementation actually moves the cursor in the
opposite direction to hop with a positive argument.

The implemented behaviour is problematic, because it allows returning to
and modifying an earlier part of the string which is likely to
invalidate saved cursor values so we're going to align the
implementation with the documentation.