ada: Flag error for string literals we don't handle
We don't currently have the special handling that would be needed to
support having non-graphic characters in string literals, so report an
error instead of quietly generating invalid Ada code.  Such characters
aren't likely to be used in Snowball's intended domain, but it's
more helpful to give an explicit error if somebody tries.

See #135.