[ci] Try partner queue a bit harder
It looks like the partner architectures only support xenial and
newer.  Not sure if we actually rely on trusty for anything, but
let's try switching to xenial by default to see.

Also enable ppc64le for now so we can see which (if any) of these
architectures actually meets our needs.