german2: Fix handling of `qu` to match description
The implementation erroneously does `skip 2` after `qu`.  We suspect
this was intended to skip the `qu` but that's already been done by the
substring/among matching, so it actually skips an extra two characters.

It seems best overall to make the code match the description of the
algorithm on the website.  This change only affects the stemming of
a single word `quae` in the sample vocabulary, which seems to actually
be Latin rather than German.