fix a memleak in snowball compiler
When an output file is used and the name option is not, then the
compiler will dynamically allocate value for the `name` option and never
free it.
This is not a large problem, because at the end of the compile process
the OS will free all allocated memory anyway.
However, when using snowball as part of a larger toolchain and then
using compile options such as `-fsanitize=leak`, a memleak in snowball
can break the entire build.
This exactly what happened to us. We could work around this somehow, but
it seems better to fix the leak in the compiler properly.

The same issue was already reported previously by a former colleague of
mine in https://github.com/snowballstem/snowball/pull/136, but that PR
seemingly was disliked by the snowball maintainers. Trying it again with
this PR with a slightly modified solution and hope it can be merged now.