Eliminate "not" from integer test code
We now generate the inverse test operator without a not for all
languages, e.g. in Python we now generate

if self.I_p1 >= self.I_x:

instead of

if not self.I_p1 < self.I_x:

This isn't going to be faster in compiled languages with an optimiser
but for scripting languages it may be faster, and even if not, it's
makes for a little less work when loading the script.