[ci] Run Python tests on full test data
The pure Python versions run slowly - when we used travis for CI
we used to need to thin the testdata for languages such as Arabic
where there's a lot to avoid the build exceeding the maximum time
allowed for a CI job.  GHA allows jobs to take up to 6 hours so
we should no longer need to do this.