[go] Dispatch among result with switch
We were generating an `if` ... `else if` chain, presumably because
the Go generator was based on the Rust generator which was based on
the Python generator, and Python didn't have a switch-like construct
until Python 3.10.

This doesn't make a measurable speed difference so it seems the
Go compiler is optimising both to equivalent code, but using a
switch here seems clearer, a better match for the intent, and is a bit
simpler to generate.