[rust] Dispatch among result with match
We were generating an `if` ... `else if` chain, presumably because
the Rust generator was based on the Python generator, and Python didn't
have a switch-like construct until Python 3.10.

Using `match` results in a 3% speed-up for an unoptimised Rust compile
but doesn't seem to make a measurable difference when optimising so
it seems the Rust compiler is optimising both to equivalent code.
However using a match here seems clearer, a better match for the intent,
and is a bit simpler to generate.