Add generic dead code elimination machinery
So far this adds a new c_functionend node and code to track what
signals can be given at points in the code.

This facilitates some new optimisations:

* Tail-calling
* Simpler code for calling routines which always give the same signal
* Simpler code when a routine ends in a integer test (this also allows
  eliminating an Ada-specific codegen optimisation which did something
  similar but only for routines which consist *entirely* of a single
  integer test.
* Dead code reporting and removal (only in simple cases currently)

Currently this overlaps in functionality with the existing reachability
tracking which is implemented on a per-language basis, and only for some
languages.  This reachability tracking was originally added for Java
where some unreachable code is invalid and result in a compile time error,
but then seems to have been copied for some other newer languages which
may or may not actually need it.  The approach it uses unfortunately
relies on correctly updating the reachability flag anywhere in the
generator code where reachability can change which has proved to be a
source of bugs, some unfixed (e.g. #83).  This new approach seems better
and with some more work should allow us to eliminate the older code.

Fixes #83