Adjust -te suffix removal rule to match docs
Due to what looks like a quirk of the implementation, we wouldn't
change -te to -t in the "doesn't have at least 4 characters before
it" case if it was preceded by a "t".  Comparing output of the
documented and implemented algorithms on a wordlist from Estonian
wikipedia, the documented version seems to do a better job overall
(and is a slightly simpler rule to describe).