Drop new warnings for and always true/or always false
On further reflection these don't seem useful - "and" on commands
has the side-effect of resetting the cursor, so `A and B` is not
redundant, and a reasonable way to express that.  An "or" where
all commands signal f is more suspect, but not likely by accident
because very few commands always signal f.