[swedish] Improve handling of -öst
Change -öst to -ös when preceded by any of 'iklnprtuv' rather than
just 'l'.

The new rule only requires the -öst to be in R1 whereas previously
we required all of -löst to be.  This second tweak doesn't seem to
affect any words ending -löst but it conflates a few extra cases
when combined with the expanded list of preceding letters, and seems
more logical linguistically (since -ös is akin to -ous in English).

Fixes #152, reported by znakeeye.