czech: Change -čté/-šté to -čtí/-ští
The Java implementation removes the latter but has incorrect comments
saying it removes the former.

Changing the Snowball implementation makes no difference here (probably
due to the oddness around when to remove a character vs calling
do_palatalise) but changing Java to use the Snowball suffixes here leads
to a clear regression, so adjust the Snowball implementation to match
Java implementation.