CzechStemmerLight: Remove one char for -es/-ém/-ím
This case was inconsistent with all the other cases where we call
palatalise as we remove the whole suffix here but leave the first
character in every over case.

Checking the vocabulary list, this means palatalise will almost never
match one of the suffixes, as the only words with this as an ending in
the list are these, which look like they're actually English words
(except "abies"):

abies
cookies
hippies
series
studies

This means palatalise will just remove the last character, which seems
odd.

This change changes a lot of stems but seems to be an improvement in
pretty much every instance I checked in google translate.