Fix handling of possessive removal
There are two issues here:

One seems clearly unintentional, which is that the cursor position from
do_case wasn't reset.

The other is that do_possessive was only called if do_case did something
which does not match the Java implementation.  It seems likely this
was not intended, and testing suggests it's not a helpful change.