We never need to restore the cursor after `not`
If `not` turns signal `f` into `t` then it sets `c` back to its old
position; otherwise, `not` signals `f` and `c` will get reset by
whatever ultimately handles this `f` (or the program exits and the
position of `c` no longer matters).

This slightly improves the generated code for the `english` and `porter`
stemmers.