[rust] Skip unnecessary calls to find_among
This change ports an optimization from the C stemmer generator to the rust generator.

In order to determine whether find_among and find_among_b are worth calling,
the generated C checks that the current string is long enough to potentially match
the shortest string literal in the among,
then does some bitwise magic to check that the characters in the current string
at length of the end of the shortest string in the among
are the same as the character at that position in at least one of the among strings.
If the current string is too short or that character doesn't match any of the amongs
at that position, then we can fail-fast instead of calling find_among.

This change allows the generated rust stemmers to similarly skip find_among calls
where we can quickly determine that find_among would not succeed.
In some stemming benchmarks I've been running, this improves the performance
of the rust english stemmer by between 40 and 50%.