greek: Fix first part of step5a to set slice
In practice it happens the slice will actually always be set as we
want, but that's only because of how the steps before this are
written so it's not good to rely on that.

This could only affect the word "αγαμε" which isn't currently in the
test vocabulary (but I'll add it shortly to provide test coverage for
this special case).