dutch.sbl: Fix UTF-8 R1 handling
The "at least 3 characters" part of the R1 definition was actually
implemented such that when working in UTF-8 it was "at least 3 bytes".

We set R1 having stripped most accents found in Dutch, but we leave `è`
alone.  I've not managed to find a Dutch word starting `è` which was
stemmed differently depending on encoding, but proper nouns and other
words of foreign origin may contain other accented characters and it
seems better for the stemmer to handle such words the same way
regardless of the encoding in use.