kraaij_pohlmann.sbl: Remove conversion of y to Y
The Snowball implementation tries to identify cases where `y` is a
consonant and temporarily changes these to `Y` which is then treated as
a consonant during stemming (then `Y` is changed back to `y` before
returning).  However the original C Kraaij-Pohlmann implementation does
not do this (it's taken from the Porter stemmers for English, French,
German and Dutch).

A quick scan of the stemming differences resulting from this change
suggests that the this extra handling only helps by conflating `royale`
with `royaal` but possibly there are additional cases and this extra
tweak is useful.  However it's getting in the way of resolving the
differences between the C and Snowball implementations so remove at
least for now and review later.

This reduces the number of words which stem differently from 65 to 45.

See #1