Fix bug in canonicalising among starter
In 811d5c18725aabe20aeb2eff8441a6daa661cced we changed to converting
an among starter to be a command before the among, adding an explicit
substring if there isn't one.  This was supposed to result in an
exactly equivalent parse tree, but we were failing to fix up the
left pointer from the among node to point to the first among case.

This was fine for C, but e.g. for Ada resulted in infinite recursion
on the Kraaij-Pohlmann implementation (which is now dutch.sbl).

Now the parse trees are exactly the same.