[C#] Workaround mcs code flow analysis bug
Put a block around each `case` in the code generated for an
`among` which seems to workaround bogus compiler errors (typically with
repeat reports at the same location):

dutchStemmer.generated.cs(543,25): error CS0165: Use of unassigned local variable `c5'

The c5 variable is initialised at point of declaration and we
don't `goto` into the block it is declared in from outside so
this seems to be buggy code flow analysis in the compiler.

Unclear where to usefully report mono bugs in 2025 so I've
not tried.