[ada] Fix `insert`, `<-` or `=` with variable
Fix code generated for `insert`, `<-` or `=`, when followed by a string
variable.  This was failing to compile with e.g.:

    stemmer-dutch.adb:421:34: error: "S_Ch" is undefined

We were missing a qualifying `Z.` and also the variable needs converting
with `Ada.Strings.Unbounded.To_String()`.

Triggered by code in the Kraaij-Pohlmann implementation (which is now
dutch.sbl).