[C#] Make csharp_stemwords -i option optional
csharp_stemwords will now read from stdin if -i is not specified,
like the C version does.

The code also no longer sucks the whole input into an array in memory,
instead reading a word and then stemming it right away.  This is more
helpful for interactive testing, and also a little faster.