python: Set minimum required Python version to 3.7
Python 2.7 reached its end of life on 2020-01-01.

Python 3.6 reached its end of life two years later, on 2021-12-23,
but that is still more than 3 years ago.

And bumping Python to 3.7 will allow me to upgrade setuptools to
61.2.0, which is the first version to properly support metadata in
pyproject.toml.

Also, replace Python 3.8 with 3.7 in the tests, so that we test both
the lower and higher bounds of supported version range.