python: No longer store Among routine name as string
If we define amongs after routines in the generated Python classes
then we can directly reference routines used in amongs by name
and avoiding a getattr() call each time an among wants to call a
routine.

This doesn't seem to make a measurable speed difference, but it's
cleaner and avoids problems with name mangling.

Suggested by David Corbett in #217.