ada,pascal: Fix reachability tracking bug
We were failing to clear our `unreachable` flag after generating
each routine, so if the end of a routine was unreachable the
snowball compiler would think the start of the next routine was
unreachable and would not generate it.  This didn't affect code
generation for any algorithms we currently ship.