Tighten parsing of writef format strings
Error out on unrecognised escape codes or if a numbered escape
is used with too high a number or a non-digit.

This change reveals that the Go and Rust generators were using invalid
escape ~A - the old writef() code was substituting this with just A
which is what is wanted so this case was harmless but being lenient
hear could hide bugs, especially when copying code between generators
as they don't all support the same set of format codes.