[ada] Fix bugs with handling of length vs limit
The Ada runtime was not tracking the current length of the string
but instead used the current limit value or some other substitute, which
manifested as various incorrect behaviours for code inside of setlimit.

`size` incorrectly returns the different between the limit and the
backwards limit.

`lenof` or `sizeof` on a string variable generated Ada code that
didn't even compile.

Some of the preconditions on these methods were also incorrect.