norwegian.sbl: Include more accented vowels
Including "ê", "ò", "ó" and "ô" improves the stemming of a fairly small
number of words, but there's basically no cost to having extra vowels in
the grouping, and a few of these words are commonly used.

We deliberately haven't included some other accented vowels which are
used in some Norwegian words:

Including "é" would conflate forms of "léta" (to paint) but "lét" is
both the imperative of "léta" and the past tense of "la" (to let/allow)
so overall this conflation doesn't seem an improvement.  It doesn't
seem to make a difference to the stemming of any other Norwegian words.

Norwegian doesn't seem to use "è" in the first vowel cluster of any
words, so including it makes not difference to the stemming of Norwegian
words.

Fixes #218, reported by András Jankovics.