hungarian.sbl: Simplify R1 definition
We were taking digraphs into account when determining where R1 starts,
to ensure it didn't start in the middle of a digraph (except that "dz"
was missing from the Snowball implementation although included in the
algorithm description).

However having R1 start in the middle of a digraph would only make a
difference to the stemming if we removed a suffix that started with
the last character of the digraph (or with "zs" in the case of "dzs").

No suffixes we remove start with "y" or "z".

Two suffixes start with "s" ("stul" and "stül") so removing 'cs' and
'dzs' makes a difference to some inputs (e.g. "acstult" and "adzstult"),
but not any inputs which are valid Hungarian words.

Removing the digraph handling speeds up stemming (by ~2% on the current
sample vocabulary list).

See #216.  Thanks to András Jankovics for confirming no Hungarian words
are affected by this change.