Go: Eliminate splitAt() helper function
We only use one half or the other in each call so it seems clearer
to just use a Go slice operation (and more efficient unless the compiler
is capable of optimising away the other unused slice operation).

It looks like this originated from the Go generator being based on the
Rust one, as Rust's string has a `split_at()` method.