Implement goto/gopast optimisation for all languages
The optimisation applies to `goto`/`gopast` applied to a grouping
or inverted grouping (which is by far the most common way to use#
`goto`/`gopast`).  It is now also implemented for Go, Java, Javascript,
Pascal and Rust.

The analyser now synthesises commands to represent these operations
which seems cleaner than checking for the situation in every backend.