Python: Generate Python with UTF-8 source encoding
This makes the generated code easier to follow, which helps during
development.  It's also a bit smaller.

For now codepoints 0x590 and above are still emitted as escape
sequences to avoid confusing rendered when LTR scripts are involved.
This is rather a crude solution, but it's just leaving these handled as
they were before.