Replace ~L0 with ~L and eliminate ~A0
The new ~L always writes p->literalstring.  In the situations where we
want to write a different string as a literal we can just call
write_literal_string() directly.

The ~A0 escape was only used (and defined) for the C generator and only
used in one place, so just replace that with a direct call to
wlitarray().