Precompute the possible signals from each command
This means we now only do this exactly once per command, whereas
previously we could end up doing it many times for some commands in some
cases.

The only functional change should we no longer make a pessimistic
assumption if the function call depth reaches 100.  This is cleaner
but is unlikely to make a difference for any real-world Snowball
programs.