C: Fix codegen for hop with constant argument
We were relying on the cursor being restored on failure by the code
which handled that failure, but if that code is a repeat or atleast
command that it has an optimisation which assumes `hop` won't do
this.  This means we generated incorrect C code for some cases where
`hop` was used inside `repeat` or `atleast`.

This doesn't functionally affect any of the stemmers we currently
ship.

Bug introduced in v2.1.0 by f480541a35a38284afe01ae88cc47c21ec92e60d.