Turn single case amongs
Sometimes a single case among is the natural way to express a single
rule in Snowball code as it can show commonality with rulesets with
multiple rules, but it's silly to actually generate as an among.

Of the stemmers we currently ship, this improves code generation for
arabic, estonian, greek and lithuanian.