Add workaround for GNU make < 4.3
With GNU make <4.3, make will fail to regenerate a manually deleted
stemmer-*.ads if the corresponding stemmer-*.adb is still present and
up-to-date.

This approach allows use to support parallel make for these GNU make
versions, and seems a reasonable trade-off.  Prior to the recent change,
our makefile wouldn't handle this case anyway, and nobody seems to have
noticed.