Fix check rules to exit on failure
Since bfc4cf3d9da850f922779ef247363d18c148958e we wouldn't exit if the
word list was gzipped (which it currently only is for arabic).

We need diff to be the last command in the if and else branches, so
the temporary file needs to be cleaned up as a separate command.