C: Implement among functions differently
The function pointers we had in struct among require relocation when the
code is in a dynamic library, which adds load-time overhead and means
the among structures can't be put in a read-only section.

Instead we now store an integer index instead, and pass in a pointer to
a dispatcher function to find_among()/find_among_b() which gets called
this index is non-zero.  The value of the index is stored in z->af so the
dispatcher function can use it.

Often only one unique function is used in an among, and in this case we
can just pass this to find_among() as the dispatcher which reduces the
overhead for this case.

Profiling with cachegrind suggests this change adds a small overhead
to algorithms which use among functions - currently finnish and hindi;
also lovins, but that's really only of academic interest and is not
enabled by default.