Go: Fix formatting of generated code
The code gets run through gofmt which was fixing up these issues,
but better to generate the code cleanly to start with.

The only things which gofmt now changes are that it indents variable
names to align in adjacent variable declarations, and a couple of things
which are apparently for compatibility with older versions of Go.