Add missing checks for x_return
The code which checks if we can combine returning on among failure
with returning at the end of the function was failing to check
(g->failure_label == x_return), which theoretically could result
in things not getting properly reinstated on failure.

In practice, it seems that currently the c_functionend marker can
only appear after an among when this additional condition is also
true, but that could change in the future if we propagated the
c_functionend marker to place where control flow on `t` next reaches
the end of the function.