Ada,C,Rust: Avoid some among helper calls
For these target languages we generate a simpler and faster check
which allows us to sometimes avoid calling the helper.

If the among only contains strings which are one byte (for UTF-8)
or one codepoint (for fixed-width encodings) and there are no
among functions used, then this faster check is actually all we
need and we don't need to ever actually call the helper for such
an among.