Python: Change slice_check() to assert conditions
In C we must not perform string slicing if slice_check() fails because
that could result in writing outside of the allocated buffer.

It's not problematic in this way for Python, and the situations
which slice_check() checks for should only happen with a Snowball
program containing logic errors, or for bugs in the Snowball compiler or
its runtime (or possibly in the Python interpreter, OS, hardware,
etc).  Therefore assert() seems an appropriate choice.

This fixes the previously buggy implementation of sliceto - an
empty string return was used to signal slice_check() failed, but
if the slice is empty the return value will be an empty string.
Luckily this case is not exercised by any current algorithms.
See #242