C#: Use Debug.Assert() in slice_check()
Previously the runtime code wrote a diagnostic message and continued
if one of these checks failed:

    System.Diagnostics.Trace.WriteLine("faulty slice operation");

The situations which slice_check() checks for should only happen with a
Snowball program containing logic errors, or for bugs in the Snowball
compiler or its runtime (or possibly in the Javascript interpreter, OS,
hardware, etc).  Therefore an assertion seems an appropriate choice,
and means the check is not enabled for a production build, which seems
more helpful overall.

See #242