Rust: Clean up return values from helpers
Previously slice_del() and slice_from() always returned true, but the
generated code included failure checks in case false was returned.
These helpers no longer return anything, and the checks are gone.

The generated code previously treated an empty string returned by
slice_to() as an error, but this was buggy since if the slice is empty
the return value will be an empty string.  The helper doesn't try
to signal an error with an empty string so we can just drop this
check.  Luckily this case is not exercised by any current algorithms.
See #242