Optimise comparing among actions
We merge among actions which are equivalent as this reduces the number
of cases in the generated code.  The node comparison function use for
this was carefully returning a full order, but actually we only need
to know if the actions are equivalent or not which can be tested more
efficiently.

For example, this reduces the time to compile serbian.sbl to C by
about 2%.