JS: Improve stemwords.js option parsing
Eliminate the check that there are at least 3 command line arguments as
we don't require any now.

If we encounter an argument we don't understand, we now report it and
show the usage message (previously we silently ignored it).

We now exit with status 1 if there's a problem parsing the command
line.