Ada: Fix string-$ problems
The code we generated for string-$ in Ada was actually partly Pascal
code (the Ada generator was originally based on the Pascal one) and
didn't even compile.

The simplest way to get this to work seems to be to represent Snowball
string variables in Ada the same way as the current string.  This means
they now take up more space (a fixed 1KB), but a typical Snowball
program has either no string variables or just one so the overhead seems
acceptable.