Ada: Don't try to adjust the slice after =S
This adjustment code is correct for `insert` and `append`, but
wrong for `=S`.  In Ada this can cause the slice ends to be
adjusted to negative values which causes a CONSTRAINT_ERROR
as the Char_Index type is defined as non-negative.

While in some situations there is a meaningful way to adjust the
slice after `=S`, in general there isn't because it can replace
part of the content of the slice.  Martin said he'd not thought
of this case, and we've concluded it's best to refine the Snowball
language definition to say the slice is unset after `=S` (as it
is defined to be initially).