Update slice end after `delete` and `<-`
The snowman said that after `[` and `]` "the slice ends will retain
the same values until altered", which doesn't make it clear what
happens for operations which modify the text the slice ends are in.

The existing handling here is inconsistent between commands: `delete`
and `<-` leave the slice ends on the same numeric positions, while
`append` and `insert` adjust the slice ends to leave the slice marking
the equivalent substring of the updated string.  When working in UTF-8
the slice end can be in the middle of a multi-byte character after
`delete` or `<-`, which seems especially undesirable.

I talked this over with Martin Porter and we've agreed that it makes
sense for `delete` and `<-` to also update the slice ends (in fact only
the right end needs adjusting) and I've clarified the snowman wording.

Existing algorithms don't seem to rely on what the slice is set to after
these commands (though it's possible that's because people developing
them ran into mysterious misbehaviour and worked around it).