PHP: Always inline code for assign_to
The code is not much longer than the call to assign_to().

Also in 3.0.0 we deprecated `=>` which is what we generate this code
for, and nothing we ship contains this command, so removing it from
BaseStemmer reduces the total code size a little.

Follows equivalent change for Javascript in
d12b77a11523f365b9a636e0a1d6f099baf0345d