PHP: Require PHP 8.3
People using a new PHP library are likely to target recent PHP versions.
If it turns out people want support for older version we can revert this
change.  Once released, it's easier to relax a version requirement than
to tighten it.

PHP 8.3 allows us to use typed class constants, which isn't vital but
is a reason to set this version as the threshold.