Include literal strings and variable names in comments
This can make the comments significantly easier to follow.

If target language comments don't go to the end of the line then
in the uncommon case of a string literal containing the end
comment sequence we use the placeholder <literal string> instead.
Of the current target languages, only C is affected by this
(because we still aim to generate C90-compatible code).

Pascal comments now use `//` rather than `{` ... `}` to avoid
this issue.  We were already using `//` comments in the Pascal
runtime.