Pascal: Eliminate While in repeat/atleast
Pascal lacks `Continue` (at least as a standard feature) and this
loop only exists so we can jump back to its start with `continue`
in other languages - we have a `Break;` at its end so it doesn't
loop in the normal way.

In Pascal we were generating a label before the loop and using
`goto` instead of `Continue`, so we can keep the label and `goto`
and scrap the loop.