Dart: Don't alias `-d` to `-dart`
Allocating a short option alias for every target language does not
seem useful enough to use up the limited supply of short options on.

Of the already supported languages, only Java has a short option alias
(`-j`) and that seems a mistake in hindsight (but removing options
is an incompatible change).