Add machinery to localise variables
If a Snowball variable is only used in one routine, that routine
doesn't call itself, and the variable is set by any code flow path which
leads to a use of the variable, then we can make that variable a local
variable instead of it being global (which typically means a private
class member in the target language).

This reduces the amount of state in stemmer objects, and typically
reduces the overhead of accessing these variables a little.

We now do this for integers and booleans in all target languages, and
for strings in target languages where benchmarking seems to show it
is faster.

The mechanism which traces the code flow errs on being too conservative
in some cases, but it's good enough for all instances in the code we
ship, and is likely to handle the vast majority of real-world cases.
We issue an "info" diagnostic to report when a variable which is only
used in one routine can't be localised - please report if you see this
in real world code and we can try to enhance the code flow tracing.