Don't convert -č to -k after suffix removal
Based on Firefox's machine translation and some looking up of words in
wiktionary, to me it seems the merges are almost all improvements, the
splits are mostly worse (but some are actually better), and the more
complicated cases are mostly better.

It seems the cases these rules help are verb and adjective forms (either
getting conflated with other forms of the same verb/adjective, or
conflating with a related noun). Perhaps there is scope for more
targetted rules which aim to address only these cases.